Backend/archive/email_outreach/requirements.txt
google-api-python-client>=2.133.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1


